Refactor domain CLI commands into separate files
Restructured domain CLI commands following the billing CLI pattern:

- Created apps/api/domains/cli/ directory structure
- Extracted DomainsHelpers module to helpers.rb
- Split domain commands into individual files:
  - list_command.rb (domains)
  - info_command.rb (domains info)
  - orphaned_command.rb (domains orphaned)
  - transfer_command.rb (domains transfer)
  - repair_command.rb (domains repair)
  - bulk_repair_command.rb (domains bulk-repair)
- Each command now uses Dry::CLI with proper registration
- Removed lib/onetime/cli/domains_command.rb
- Updated lib/onetime/cli.rb to remove explicit require
- Commands are now auto-discovered by the CLI loader

This follows the established pattern used by billing commands in
apps/web/billing/cli/ and improves code organization and maintainability.
