+++ lib/onetime/cli.rb
@@ -67,7 +67,6 @@ require_relative 'cli/migrate_redis_data_command'
 require_relative 'cli/sync_auth_accounts_command'
 require_relative 'cli/customers_command'
 require_relative 'cli/test_data_command'
-require_relative 'cli/domains_command'
 require_relative 'cli/billing_command'
 require_relative 'cli/change_email_command'
 require_relative 'cli/session_command'
